[IMP] convert to native javascript modules
